+++ src/Cshrix/Serialization/UnixMillisecondDateTimeConverter.cs
@@ -33,10 +33,10 @@ namespace Cshrix.Serialization
 
     using Extensions;
 
-    using Helpers;
-
     using Newtonsoft.Json;
     using Newtonsoft.Json.Converters;
+
+    using Utilities;
 
     /// <inheritdoc />
     /// <summary>
@@ -90,7 +90,7 @@ namespace Cshrix.Serialization
             object existingValue,
             JsonSerializer serializer)
         {
-            var nullable = ReflectionHelpers.IsNullable(objectType);
+            var nullable = ReflectionUtils.IsNullable(objectType);
 
             if (reader.TokenType == JsonToken.Null)
             {
