+++ src/Cshrix/Utilities/ReflectionUtils.cs
@@ -6,14 +6,14 @@
 //   file, You can obtain one at http://mozilla.org/MPL/2.0/.
 // </copyright>
 
-namespace Cshrix.Helpers
+namespace Cshrix.Utilities
 {
     using System;
 
     /// <summary>
     /// Contains helper methods for performing reflection tasks.
     /// </summary>
-    internal static class ReflectionHelpers
+    internal static class ReflectionUtils
     {
         /// <summary>
         /// Checks if a type is not a value type and nullable.
