+++ src/Cshrix/Utilities/TransactionUtils.cs
@@ -6,7 +6,7 @@
 //   file, You can obtain one at http://mozilla.org/MPL/2.0/.
 // </copyright>
 
-namespace Cshrix.Helpers
+namespace Cshrix.Utilities
 {
     using System;
     using System.Threading;
